enable submission detailview in submission table
